changed directory for code to src namespacing now uses static::class instead of strings updated config.yml
